Added cleanup / publish scripts

+++ scripts/publish.js
@@ -10,8 +10,12 @@ function prepareNpmPublish() {
   let version = json.version.split(".");
   version[version.length - 1] = Number(version[version.length - 1]) + 1;
   json.version = version.join(".");
-  fs.writeFileSync("publish/package.json", JSON.stringify(json, null, 2) ,"utf8");
-  console.log("done !");
+  if (!fs.existsSync("publish/package.json")) {
+    fs.writeFileSync("publish/package.json", JSON.stringify(json, null, 2) ,"utf8");
+    console.log("done !");
+  } else {
+    console.error("out / publish directory not found. Please run yarn build first")
+  }
 }
 
 prepareNpmPublish();
